+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Version 0.2.0
 
 ### Added
+* Add `Regions.get_item`
 * GDAL COG-backend for CocoFrames
 * `CocoDataset.remove_images`